* MonoDevelop.Projects/Project.cs: Don't delete files from the   output directory if they were directly referenced from that   directory. Fixes bug #472620 - Doing a Clean or Rebuild removes all   referenced assemblies from output dir even if they aren't built.
svn path=/trunk/monodevelop/; revision=126821

+++ main/src/core/MonoDevelop.Projects/MonoDevelop.Projects/Project.cs
@@ -359,6 +359,10 @@ namespace MonoDevelop.Projects
 			foreach (FileCopySet.Item item in GetSupportFileList (solutionConfiguration)) {
 				string dest = Path.Combine (config.OutputDirectory, item.Target);
 				
+				// Ignore files which were not copied
+				if (Path.GetFullPath (dest) == Path.GetFullPath (item.Src))
+					continue;
+				
 				try {
 					if (File.Exists (dest)) {
 						FileService.DeleteFile (dest);
